Optionally deal with missing lookup entries by calling an external script.
This adds a configuration option, externalLookupHelper, that contains the path
to a script that, when called with a MAC address, will report the IP address
that MAC currently has, if any.

The intent is to deal with situations where RackHD is not running on the DHCP
server and so can't be guaranteed to have either the ARP table entries or
the DHCP server's lease entries to find lookup entry data. In this case the
script can be used to provide the necessary data.